Changed target Fx to .NET 4.5,
because this will enable us to target operations on finite sequences (such
as PickRandom, ToSet, ToQueue, etc.) on IReadOnlyCollection<T> instead of
IEnumerable<T>.

The original motivation for targeting .NET 3.5 was to make Champagne as
generally useful as possible. This should still be a goal, but will
probably mean that we'd need to look into conditional compilation at a
later date, in order to make a special build for .NET 3.5.

Perhaps other users also need a version for Windows Phone, etc...